+++ webapp/forms.py
@@ -21,6 +21,6 @@ class InvestmentForm(Form):
 class MonteCarloForm(Form):
     rate = DecimalField('Rate', places=2)
     sigma = DecimalField('Sigma', places=2)
-    time = IntegerField('Years')
+    time = DecimalField('Years')
     start_val = DecimalField('Initial Value', places=2)
     annual_contrib = DecimalField('Annual Contribution', places=2)
